add details. test=develop
